extract welcome window into a component to display when disconnected from polkadot in the account page
